+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+Next Release
+------------
+Nothing yet.
+
+Release 0.126.1
+---------------
+- Validate role expiration time for permission granting user.
+- Fix settings name update through service name update.
+
 Release 0.126.0
 ---------------
 - Migrate to Django 1.9